Address feedback
- Refactor away from dynamic dispatch for object storage
- Replace file buffering with streaming to storage
- Add proper handling of missing evidence submission on attachment upload
- More repository cleanup
- Use usize for max uploaded bytes
- Clean up tests

+++ src/config/mod.rs
@@ -77,7 +77,7 @@ pub struct GcsObjectStorageConfig {
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct UploadsConfig {
-    pub max_attachment_bytes: u64,
+    pub max_attachment_bytes: usize,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
